Additional headers and properties support
#323 Added support for additional headers and properties support

The additional headers will be added on top of the existing headers
from the message.

The properties added are only basic AMQP properties as defined in the
{@link com.rabbitmq.client.AMQP.BasicProperties.Builder}. When the
message contains already these properties then will be considered and
these additional properties will be ignored.

Made RabbitMQConstants to enum and added the documentation from docs.
ToDo: The documentation needs to be generated from the source code.